Our builds stopped working because latest is no longer working
This PR switches us to using glide to build the plugin and pins on specific versions
of the libraries that we require for the plugin.

We'll also change the Dockerfile build process to use the current source for our
builds vs trying to git clone it from the remote repos.

Signed-off-by: Edward Raigosa <edward.raigosa@hpe.com>